feat: add locations API with country, state, and city dropdowns
